refactorización para limpieza de código
